Also generate a copy with relative paths only
This is useful for demonstration, as the resulting website can be copied
around and opened from other locations, even the local file system, without
being too much damaged.  The heuristics used to alter the code are pretty
simple, though, so this might have adverse effects in some corner cases and
is not intended to be used for the official website.  Hence the different
directory name.