Link to ICMS 2016 presentations and extended abstracts
The extended abstracts themselves are in a separate directory (/docs) and
not part of this commit, so that they don't end up in the repository since
the self-archival conditions from Springer don't extend to GitHub storage.

The pub/index file is just a minimalistic directory listing, with no links
pointing to it.  In its current form it is not meant to be viewed by users,
but those who manually edit the URL in the address bar will now receive this
managed document instead of a server-generated directory listing.  Once we
have multiple publications, that page should probably contain more
information, in a format yet to be decided.