Fix links to GitHub
The links for “View source” or “Edit page” in the bottom right corner of
e.g. raytracer.html are currently broken, lacking the “src” segment of the
path.  In order to fix this, the change at hand avoids the use of the cwd
option to gulp.src.  Instead the path is included in the patterns, and the
base option will see that it gets stripped.  That way the github source
links contain the full path information and work as intended.