Add GitHub button to topbar
The button uses only the GitHub mark, with no button text.  That way people
who know GitHub will quickly find their way, without wasting too much space
for those who don't know what to make of the name anyway.

https://github.com/logos suggests using the GitHub mark for the purpose of
linking to the GitHub project page, as opposed to the textual logo or the
full-color octocat, both of which are intended to signify GitHub as a whole.
The SVG was chosen in preference over the pixel-based icons available from
that page since we don't use a pixel-based layout for our top bar.