Adjust ICMS paper links to extras
Using /docs to contain documents outside the repository was a bad idea,
since that path was already taken for the documentation, leading to
conflicts.  Now the extra files are located in /extras server-side, and
although there is a redirect in place it would be prudent to adjust the
links.  Adjusting the gulpfile allows testing the documentation, too.