Format reference manual more like its stand-alone version
This imports and adjusts a number of settings from CindyJS/ref/ref.css, so
that the manual developers can create locally and the one on the web page
look more alike.  If the web version should change in the future, it might
make sense to port these changes to CindyJS for the sake of consistency.

Some of the formatting, like e.g. the default format for tables, will
affect other pages as well.  The modifications to the table of content were
a drive-by fix, motivated by the obervation that code block background and
floated toc don't work well together.  And many toc rows will be too long
for a float anyway, so let's just consistently prepend the toc to the actual
page content.