Improved footer layout
The foters are now aligned at their top, and space gets distributed around
them without using an auto margin for that purpose.  This makes layout more
consistent and easier to control when switching to small-screen mode.  There
we go back from flex to block display.  This was present in the past but got
broken at some point when the nested div got introduced.