Switch to vertical example layout at larger sizes
The control elements included in the text of the CindyGL example on the
start page cause horizontal scrolling on some browsers, which we should
avoid.  And even without horizontal scrolling, the text would be VERY tall
and slim, looking hardly aesthetic.  So now we switch to vertical layout
(text below image instead of on the side of it) at larger sizes.  And we
center the widgets in that layout mode, which makes it look way better than
it used to.