Improved support for block quotes and syntax highlighting
I'm using this for Sage code, as the Jupyter Sage UI can export to markdown
and will enable Python highlighting for its code snippets.

The white shadow behind an image that is included in a block quote will make
this image blend in more seamlessly without requiring a white background for
the whole blockquote.